Split more screens out of RN core
