+++ apps/native-component-list/src/screens/ReactNativeCore/ReactNativeCoreScreen.web.tsx
@@ -14,9 +14,6 @@ import {
 
 import Colors from '../../constants/Colors';
 import Layout from '../../constants/Layout';
-import { AlertExample } from './Alert';
-import { ClipboardExample } from './Clipboard';
-import { StatusBarExample } from './StatusBar';
 
 export default class ReactNativeCoreScreen extends React.Component {
   state = {
@@ -38,9 +35,6 @@ export default class ReactNativeCoreScreen extends React.Component {
         removeClippedSubviews={false}
         contentContainerStyle={{ backgroundColor: '#fff' }}
         sections={[
-          { title: 'StatusBar', data: [() => <StatusBarExample />] },
-          { title: 'Alert', data: [() => <AlertExample />] },
-          { title: 'Clipboard', data: [() => <ClipboardExample />] },
           { title: 'Vertical ScrollView, RefreshControl', data: [this._renderRefreshControl] },
           { title: 'Horizontal ScrollView', data: [this._renderHorizontalScrollView] },
           { title: 'Text', data: [this._renderText] },
